Rename mention_command.py to markdown.py
Use a more general name for this file in utils, as it will allow it to
potentially also contain some other functions / functionalities in the
future.

Having single-purpose files named the same as the function that they
contain is an overkill in this case and would likely soon result in a
lot of clutter in the utils/ directory.

Additionally, having the file named the same as a function, that is then
re-exported in `__init__.py` causes some issues if someone would want to
import this file explicitly as a module, since the re-exported function
would take precedence.