Add docker-compose that builds the image locally
Currently, the project only contains a docker-compose file that relies
on ghcr images, which are released automatically using a CI action when
an update comes out in the project's `main` branch.

However, in some situations, it can be useful to allow people to quickly
use docker-compose that will build the image for them locally, to allow
for testing something in a system agnostic way within the container.